todo ok menos hex

+++ srcs/ft_printf_select_var.c
@@ -33,9 +33,9 @@ void	ft_var_print_main(char c, va_list args, int *sol, t_options *flags)
 	else if (c == 'u' && flags->minus == 0)
 		ft_var_print_unsigned_right(va_arg(args, unsigned int), sol, flags);
 	else if ((c == 'x' || c == 'X') && flags->minus == 1)
-		ft_var_print_hex_left(c, va_arg(args, unsigned int), sol, flags);
+		ft_var_print_hex_left(c, va_arg(args, long int), sol, flags);
 	else if ((c == 'x' || c == 'X') && flags->minus == 0)
-		ft_var_print_hex_right(c, va_arg(args, unsigned int), sol, flags);
+		ft_var_print_hex_right(c, va_arg(args, long int), sol, flags);
 	else
 		*sol = -1;
 }
